fix service Docker image
